requirements.txt
PyQt6
openai-whisper>=20231117
faster-whisper>=1.0.0
torch>=2.0.0
torch>=2.0.0
google-generativeai>=0.3.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0